rifts_character_creator/requirements.txt
boto3>=1.22.11
django-cors-headers>=3.10.0
django>4.0
djangorestframework>=3.13.1
psycopg2-binary>=2.9.3
psycopg2>=2.9.3
zappa==0.58.0


